arm64: Select ARCH_HAS_FAST_MULTIPLIER
It is probably safe to assume that all Armv8-A implementations have a
multiplier whose efficiency is comparable or better than a sequence of
three or so register-dependent arithmetic instructions. Select
ARCH_HAS_FAST_MULTIPLIER to get ever-so-slightly nicer codegen in the
few dusty old corners which care.

In a contrived benchmark calling hweight64() in a loop, this does indeed
turn out to be a small win overall, with no measurable impact on
Cortex-A57 but about 5% performance improvement on Cortex-A53.

Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: sohamxda7 <sensoham135@gmail.com>
Signed-off-by: Kneba <abenkenary3@gmail.com>
Signed-off-by: Tiktodz <ewprjkt@proton.me>